DOCERIA BACKEND/requirements.txt
fastapi
uvicorn
python-jose
passlib[bcrypt]
sqlalchemy
pydantic
pydantic-settings
python-multipart
email-validator
